+++ CHANGES.md
@@ -1,8 +1,18 @@
 # CHANGES
 
+## 2025/09/12
+
+- Extension to arbitrary deep tree structures.
+
+- New localStorage *info* object.
+
+- Real-time access to checked nodes via *info* object.
+
+
 ## 2025/09/06
 
 - Fix on *localstoragekey* option.
+
 
 ## 2025/08/22